revert the important parts of this previous commit, from 21 May:
 > eliminate the double refresh that happens when you're focused on the
 > GPS, but it's not in the center, and you zoom the screen.  previously
 > the screen would zoom around the center, and then the next GPS movement
 > (which might not be immediate) would readjust the offset from
 > center.  now everything gets lined up right the first time.

the trouble arises when there's no GPS present (either because it has
gone away, or, more importantly, because it never existed).  in this
case, zooming will continue following the GPS, rather than the center
of the screen, because there's no callback from the (missing) GPS to
tell us not to follow the GPS anymore.

so the double refresh when zooming, while the GPS is present but
off-center, is back.