buildmap_osm_text speedup
i'm not sure what the broader ramifications of this change are, but
i'm now able to convert the massachusetts-latest.osm file from
geofabrik in less than 10 hours -- it ran for a day and a half
originally before i killed it.

mainly i eliminated extra searches, and the .rdm file may end
up with more nodes or ways than it would have originally, but
i'm not even convinced of that.

there's a good chance that this has broken code that was meant
to deal with the "splitter" output.  i have no way of testing that
at present.