buildmap_osm: can now create quadtiles using the native XML parser
previously we relied on the OSM Binary Protocol to produce quadtiles.
the trouble was that that protocol never took off, and there are no
public servers for that protocol.  so we used a (slow) PHP script to
translate XML from the regular servers into the binary protocol.  this
commit eliminates the need for that script.