increase the maximum polygon node count to a 32 bit value
initially, there could only be a total of 65535 polygon nodes in a
database.  (well, twice that, because both the offset into the table
of node ids, as well as the count of node ids, were each 16 bits.)
these limits were later extended to 20 bits by reclaiming an unused
filler byte, giving 4 bits more for the offset and the count.

this commit gives 32 bits to to each, costing every polygon another 4
bytes.  we preserve backward compatibility by detecting, when the
database is opened, which version we have by looking at the total
length of the polygon data vs. the count of polygon structures.  so
this code can use old maps, but old code can't read new maps.

buildmap will no longer produce the old format.