gtk2/roadmap_canvas.c: fix label positions
this involves a bunch of trickery to get the rotated labels into
the right position on the screen.  if we switched to using a
cairo surface rather than a pixmap to draw the image on, i think
we could position the label just where we want it without the magic.