roadmap_trip: consolidate some more HAVE_NAVIGATE_PLUGIN ifdefs
if or when we start working on navigation again, we need to fully
separate it from the current roadmap_trip implementation.  for better
or worse, the roadmap_trip code is all about following GPX.  the
navigate code will be something quite different, and should remain
separate until we see what can be merged safely.