roadmap_locator:  create mapping between internal layers and map layers
this mapping will allow breaking the current hard dependency that
roadmap must be run with the same class file that was used to build
the maps.

it's very arguable that this mapping shouldn't live in roadmap_locator.
but the results should be cacheable, and the mapping needs to happen
after the map's metadata module is available.  so i'm leaving it here
for now.  perhaps someday it could move to a dummy database module
whose activate() routine establishes the mapping?